added ability to update person with ensemble
Created the foundational functionality to be able to pass arguments to the Person's ensemble. I did this by giving the Person model an ensemble array that will accept the Ensemble type objectId as well as an ensemble name to be stored.

A person should be able to have many ensembles. This functionality will be applied to the sections as well.