app refactor see description
We are refactoring the hierarchy of the data tree to now include a Program. A program would be defined as an organizations choir, band etc... type of music program.

We've adjusted account creation to now also create a program simultaneously. When data is queried, it will be queried from the Program instead of the user. This will allow for scalability in the future if the user may have more than one program.

This move opens us up to school wide accounts and possibly district wide accounts